backend/requirements.txt
fastapi
uvicorn
python-multipart
transformers
huggingface_hub
torch
accelerate
python-dotenv
PyPDF2
